VFS integration test
