+++ Makefile
@@ -34,7 +34,7 @@ endif
 .PHONY: vfs
 vfs:
 	mkdir -p dist
-	go build -tags SQLITE3VFS_LOADABLE_EXT,sqlite_os_trace -o dist/litestream-vfs.a -buildmode=c-archive cmd/litestream-vfs/litestream-vfs.go
+	go build -tags SQLITE3VFS_LOADABLE_EXT -o dist/litestream-vfs.a -buildmode=c-archive cmd/litestream-vfs/litestream-vfs.go
 	mv dist/litestream-vfs.h src/litestream-vfs.h
 	gcc -framework CoreFoundation -framework Security -lresolv -g -fPIC -shared -o dist/litestream-vfs.so src/litestream-vfs.c dist/litestream-vfs.a
